django full course

+++ full_course/main/views.py
@@ -1,10 +1,34 @@
 from django.shortcuts import render, redirect
 from .models import RoomDatabase, Topic
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.http import HttpResponse
 from django.db.models import Q
 from .forms import RoomForm
 # Create your views here.
 
 
+def login_page(request):
+    if request.method == "POST":
+        username = request.POST["username"]
+        password = request.POST["password"]
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('room')
+        else:
+            messages.error(request, 'username or password does not exist')
+    return render(request, 'home/login_reg.html')
+
+
+def logout_page(request):
+    logout(request)
+    return redirect('room')
+
+
+@login_required(login_url='login')
 def room(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = RoomDatabase.objects.filter(
@@ -38,9 +62,12 @@ def create_room(request):
     return render(request, 'home/room_form.html', {"forms": forms})
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = RoomDatabase.objects.get(id=pk)
     forms = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("you are not allow here")
     if request.method == "POST":
         forms = RoomForm(request.POST, instance=room)
         if forms.is_valid():
@@ -49,8 +76,11 @@ def update_room(request, pk):
     return render(request, 'home/room_form.html', {"forms": forms})
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = RoomDatabase.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("you are not allow here")
     if request.method == "POST":
         room.delete()
         return redirect('room')
